Store a symbol's defining subsection as a u32 index
Symbol.isec was Option<InputSectionId> (Option<usize>, 16 bytes).
Subsection indices fit comfortably in u32, so it becomes Option<u32>
(8 bytes); call sites map it back to usize where they index ctx.isecs.
Symbol drops from 72 to 64 bytes. Output unchanged; 322 tests pass.

+++ src/thunks.rs
@@ -141,7 +141,7 @@ fn scan_relocs_into_thunk<E: Arch>(
 
             let sym = &ctx.symtab[sym_id];
             if let (Origin::Obj(_), Some(target)) = (sym.origin, sym.isec) {
-                let t = &ctx.isecs[ctx.resolve_isec(target)];
+                let t = &ctx.isecs[ctx.resolve_isec(target as usize)];
                 // A target in another output section has no offset in
                 // this section's space; reserve an entry.
                 if t.osec != osec {
